pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gitpython>=3.1.44",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "flask-login>=0.6.3",
    "pyyaml>=6.0.2",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "tomli>=2.2.1",
]
